avoid sklern.scale() error for converting float32 to float64

+++ classifier/preprocessing/data_preprocessing.py
@@ -68,7 +68,7 @@ def standardization(dataset):
 
 	print(bcolors.YELLOW + 'Running standardization' + bcolors.ENDC)
 	# Standardization
-	scaledDataset = preprocessing.scale(dataset)
+	scaledDataset = preprocessing.scale(np.float64(dataset))
 
 	return scaledDataset
